+++ asilib/asi/trex.py
@@ -28,6 +28,8 @@ import numpy as np
 import pandas as pd
 import scipy.io
 import requests
+import matplotlib.pyplot as plt
+import matplotlib.colors
 
 import asilib
 from asilib.asi.themis import _get_pgm_files
@@ -517,6 +519,120 @@ def trex_rgb_info() -> pd.DataFrame:
     df = df[df['array'] == 'TREx_RGB']
     return df.reset_index(drop=True)
 
+
+def trex_rgb_available(time: utils._time_type = None, time_range: utils._time_range_type = None) -> pd.DataFrame:
+    """
+    Check for TREx RGB ASI data availability online for a given time or in a given time range.
+
+    Currently the availability is reported to within an hour.
+
+    Parameters
+    ----------
+    time: str or datetime.datetime
+        A time to look for the ASI data at. Either time or time_range
+        must be specified (not both or neither).
+    time_range: list of str or datetime.datetime
+        A length 2 list of string-formatted times or datetimes to bracket
+        the ASI data time interval.
+    """
+
+    if (time is None) and (time_range is None):
+        raise ValueError('time or time_range must be specified.')
+    elif (time is not None) and (time_range is not None):
+        raise ValueError('both time and time_range can not be simultaneously specified.')
+
+    base_url = rgb_base_url
+
+    if time is not None:
+        time = utils.validate_time(time)
+        _start = time.replace(minute=0, second=0, microsecond=0)
+        time_range = (_start, _start + timedelta(hours=1))
+
+    time_range = utils.validate_time_range(time_range)
+    start = time_range[0].replace(minute=0, second=0, microsecond=0)
+    days = pd.date_range(start, time_range[1], freq='D')
+
+    hours = pd.date_range(start, time_range[1], freq='h')
+    _available_df = pd.DataFrame(index=hours, columns=trex_rgb_info()['location_code'])
+    _available_df[:] = False
+    _current_date = datetime.min.date()
+
+    if days.shape[0] > 1:
+        days = utils.progressbar(days, text='Checking TREx RGB availability')
+
+    for day in days:
+        if _current_date != day.date():
+            start_url = base_url + f'{day.year}/{day.month:02}/{day.day:02}/'
+            _current_date = day.date()
+
+            try:
+                d = download.Downloader(start_url, headers={'User-Agent':'asilib'})
+                daily_folders = d.ls(f'*_rgb*')
+            except requests.exceptions.HTTPError as err:
+                if '404 Client Error: Not Found for url:' in str(err):
+                    continue
+                raise            
+
+        for daily_folder in daily_folders:
+            location_code = daily_folder.url.split('/')[-2].split('_')[0]
+            hour_folders = daily_folder.ls('ut*/')
+            dates_hours_on = [
+                datetime(
+                    day.year,
+                    day.month,
+                    day.day,
+                    int(hour_folder.url.split("/")[-2][2:])
+                )
+                for hour_folder in hour_folders
+            ]
+            for date_hour_on in dates_hours_on:
+                if (date_hour_on >= time_range[0]) and (date_hour_on < time_range[1]):
+                    _available_df.loc[date_hour_on, location_code.upper()] = True
+    _available_df[pd.isna(_available_df)] = False
+
+    if time is not None:
+        _available_df = _available_df.loc[time_range[0], :]
+    return _available_df
+
+
+def plot_trex_rgb_available(time_range=None, ax=None, base_url=None, pcolormesh_kwargs={}):
+    """
+    Plot the availability of TREx RGB ASI data online for a given time range.
+
+    Parameters
+    ----------
+    time_range: list of str or datetime.datetime
+        The ASI data time interval.
+    ax: matplotlib.axes.Axes, optional
+        The axes to plot on. If None, a new figure and axes are created.
+    base_url: str
+        Not used (kept for API parity with THEMIS helper).
+    pcolormesh_kwargs: dict
+        Additional keyword arguments to pass to `plt.pcolormesh`.
+
+    Returns
+    -------
+    tuple
+        A tuple containing the availability DataFrame and the axes object.
+    """
+    available_df = trex_rgb_available(time_range=time_range)
+    if ax is None:
+        fig, ax = plt.subplots(figsize=(10, 5))
+
+    pcolormesh_kwargs['cmap'] = pcolormesh_kwargs.get(
+        'cmap',
+        matplotlib.colors.ListedColormap(['lightgray', 'lightgreen'])
+    )
+    pcolormesh_kwargs['edgecolors'] = pcolormesh_kwargs.get('edgecolors', 'black')
+    pcolormesh_kwargs['linewidth'] = pcolormesh_kwargs.get('linewidth', 0.5)
+
+    ax.pcolormesh(
+        available_df.index,
+        available_df.columns,
+        available_df.to_numpy().astype(bool).T,
+        **pcolormesh_kwargs
+    )
+    return available_df, ax
 
 def trex_nir(
     location_code: str,
@@ -1779,3 +1895,20 @@ def read_nir(file_list, workers=1, first_frame=False, no_metadata=False, quiet=F
     # return
     data = None
     return images, metadata_dict_list, problematic_file_list
+
+if __name__ == "__main__":
+    # a = themis_available(time='2018-01-31T14:00')
+    import time
+    print('Sleeping...')
+    time.sleep(2)
+    # time_range = ('2018-01-28T00:01:27', '2018-01-31T23:00')
+    # # availability = themis_available(time_range=time_range)
+    # # print(a)
+    # pass
+    # plot_themis_availability(time_range=time_range)
+    # plt.show()
+
+    time_range = ('2021-11-01T00:00', '2021-11-05T00:00')
+    plot_trex_rgb_available(time_range=time_range)
+    plt.show()
+    pass